feat: Modify main for simple components test
This commit changes `os_init` to only initialize the deviceManager
(IMU, Display, Buttons and DRV). No WiFi connection is established.
After initialization, 4 tasks are created:
- update the Display
- check button events
- read accelerometer data and print to serial monitor
- play first DRV effect

NOTE: No error checking is done in the event of a miss configuration.
Check the serial monitor for error logs in that case (the errors
are not useful for knowing exactly what happened, but they at least
tell you what failed).